Set up buildbot roles and vagrant box
This isn't used live, not is it confirmed to be working in its current
state, but I'd like to merge it now, for posterity's sake.

We may end up deleting the code later (or hopefully fixing it up!),
but I'd like to close this PR that's been sitting around.
The code is isolated to its own custom role(s), so it shouldn't affect
anything else.

- [1] https://github.com/Stanford-Online/configuration/pull/166

Rebased-by: stv <stv@stanford.edu>